Factorize common code into env.cs module
