+++ dotnet/dotNetResponse.cs
@@ -15,8 +15,6 @@
 */
 
 using System;
-using System.IO;
-using System.Collections.Generic;
 using Newtonsoft.Json;
 using IBM.XMS;
 
@@ -25,8 +23,7 @@ namespace ibmmq_samples
 {
     class SimpleResponse
     {
-        private ConnVariables conn = null;
-        private MQEndPoints points = null;
+        private Env env = new Env();
 
         private const int TIMEOUTTIME = 30000;
         private static bool keepRunning = true;
@@ -43,7 +40,7 @@ namespace ibmmq_samples
             try
             {
                 SimpleResponse simpleConsumer = new SimpleResponse();
-                if (simpleConsumer.EnvironmentIsSet())
+                if (simpleConsumer.env.EnvironmentIsSet())
                     simpleConsumer.ReceiveMessages();
             }
             catch (XMSException ex)
@@ -72,6 +69,8 @@ namespace ibmmq_samples
             IDestination destination;
             IMessageConsumer consumer;
             ITextMessage textMessage;
+
+            Env.ConnVariables conn = env.Conn;
 
             // Get an instance of factory.
             factoryFactory = XMSFactoryFactory.GetInstance(XMSC.CT_WMQ);
@@ -169,43 +168,6 @@ namespace ibmmq_samples
             }
         }
 
-        bool EnvironmentIsSet()
-        {
-            bool isSet = false;
-            try
-            {
-                Console.WriteLine("Looking for file");
-                using (StreamReader r = new StreamReader("env.json"))
-                {
-                    Console.WriteLine("File found");
-                    string json = r.ReadToEnd();
-
-                    points = JsonConvert.DeserializeObject<MQEndPoints>(json);
-
-                    if (points != null && points.mq_endpoints != null && points.mq_endpoints.Count > 0)
-                    {
-                        conn = points.mq_endpoints[0];
-                        conn.dump();
-                        isSet = true;
-                    }
-                    else
-                    {
-                        Console.WriteLine("MQ Settings not found, unable to determine connection variables");
-                    }
-                    Console.WriteLine("");
-                }
-                return isSet;
-            }
-            catch (Exception e)
-            {
-                Console.WriteLine("Exception caught: {0}", e);
-                Console.WriteLine(e.GetBaseException());
-                return isSet;
-            }
-
-        }
-
-
         public class JsonMessage
         {
             public string msg;
@@ -228,36 +190,6 @@ namespace ibmmq_samples
             public string toJsonString()
             {
                 return JsonConvert.SerializeObject(this);
-            }
-        }
-
-        public class MQEndPoints
-        {
-            public List<ConnVariables> mq_endpoints;
-        }
-
-        public class ConnVariables
-        {
-            public string host;
-            public string qmgr;
-            public int port;
-            public string channel;
-            public string queue_name;
-            public string app_user;
-            public string app_password;
-            public string cipher_suite;
-            public string key_repository;
-            public void dump()
-            {
-                Console.WriteLine("hostname{0} ", host);
-                Console.WriteLine("port{0} ", port);
-                Console.WriteLine("qmgr{0} ", qmgr);
-                Console.WriteLine("channel{0} ", channel);
-                Console.WriteLine("queue{0} ", queue_name);
-                Console.WriteLine("app_user{0} ", app_user);
-                //Console.WriteLine("app_password{0} ", app_password);
-                Console.WriteLine("cipherSpec{0} ", cipher_suite);
-                Console.WriteLine("sslKeyRepository{0} ", key_repository);
             }
         }
     }
